+++ app/components/RemoveBtnThread.jsx
@@ -1,15 +1,16 @@
 "use client";
-
 import { useRouter } from "next/navigation";
-//import { HiOutlineTrash } from "react-icons/hi";
 import { FaTrashAlt } from "react-icons/fa";
+import { UserAuth } from "../context/AuthContext";
 
-export default function RemoveBtn({ id }) {
+export default function RemoveBtn({ id, userId }) {
+  const { user } = UserAuth();
   const router = useRouter();
+  const userUID = user && user.uid;
   const removeThread = async () => {
     const confirmed = confirm("are your sure ?");
 
-    if (confirmed) {
+    if (confirmed && userId === userUID) {
       const response = await fetch(`/api/threads?id=${id}`, {
         method: "DELETE",
       });
@@ -17,11 +18,16 @@ export default function RemoveBtn({ id }) {
         alert("Thread deleted");
         router.refresh();
       }
+    } else {
+      alert("ce n'est pas ton post");
+      router.push("/login");
     }
   };
   return (
-    <button className="remove-thread" onClick={removeThread}>
-      <FaTrashAlt  />
-    </button>
+    <>
+      <button className="remove-thread" onClick={removeThread}>
+        <FaTrashAlt />
+      </button>
+    </>
   );
 }
